+++ resources/sql/queries.sql
@@ -11,12 +11,12 @@ SET password = :password
 WHERE login= :login
 
 -- :name get-user :? :1
--- :doc retrieves a user record given the id
+-- :doc retrieves a user record given the login
 SELECT * FROM users
 WHERE login = :login
 
 -- :name delete-user! :! :n
--- :doc deletes a user record given the id
+-- :doc deletes a user record given the login
 DELETE FROM users
 WHERE login = :login
 
@@ -40,13 +40,37 @@ WHERE id = :id
 -- :name get-problem :? :1
 -- :doc retrieves a problem record given the num
 SELECT * FROM problems
-WHERE id = :id
+WHERE num = :num
 
 -- :name delete-problem! :! :n
--- :doc deletes a problem record given the num
+-- :doc deletes a problem record given the id
 DELETE FROM problems
 WHERE id = :id
 
 -- :name problems :? :*
 -- :doc get all problems
 SELECT * from problems order by num
+
+-- :name delete-problems-all! :! :n
+-- :doc delete all from problems table
+DELETE FROM problems
+
+-- :name problems-count :? :1
+-- :doc returns problems count
+SELECT COUNT(*) FROM problems
+
+-- :name create-answer! :! :n
+-- :doc creates a new problem record
+INSERT INTO answers
+(login, num, answer, md5)
+VALUES (:login, :num, :answer, :md5)
+
+-- :name get-answer :? :1
+-- :doc retrieves an answer to `num` from user `login`
+SELECT * FROM answers
+WHERE num = :num and login = :login
+
+-- :name answers :? :*
+-- :doc retrive answers solved by user `login`
+SELECT * from answers
+WHERE login = :login